requirements.txt
spotipy
pandas
python-dotenv
sqlite3-binary